+++ qase-cucumberjs/changelog.md
@@ -1,3 +1,10 @@
+# qase-cucumberjs@2.1.5
+
+## What's new
+
+- Added support for status filter in the test run.
+- Improved error handling.
+
 # qase-cucumberjs@2.1.3
 
 ## What's new